+++ MOP/src/Common/ExceptionManager.cs
@@ -142,7 +142,8 @@ namespace MOP.Common
             output += $"Date and Time: {DateTime.Now:yyyy-MM-ddTHH:mm:ssZ}\n";
             output += $"{GetSystemInfo()}\n";
             output += $"Session ID: {MOP.SessionID}\n";
-            output += $"Screen resolution: {Screen.width}x{Screen.height}\n\n";
+            output += $"Game resolution: {Screen.width}x{Screen.height}\n";
+            output += $"Screen resolution: {Screen.currentResolution.width}x{Screen.currentResolution.height}\n\n";
 
             output += "=== MOP SETTINGS ===\n\n";
             output += $"ActiveDistance: {MopSettings.ActiveDistance}\n";