+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - MOP now comes with two variants! A version with the legacy Mod Loader, and a shiny new Mod Loader Pro support!
   - In order to fully utilize the new Mod Loader Pro, download "MOP Pro" file
+- (Mod Loader Pro) Added "Change Resolution" button in MOP settings
 
 ### Changes